requirements.txt
click==8.0.3
pydantic==1.8.2
PyYAML==6.0
typer==0.4.0
typing-extensions==4.0.1
